refactor: hooks and components
