+++ client/src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Board } from "./Board"
+export { VictoryBanner } from "./VictoryBanner"
+export { BoardResizer } from "./BoardResizer"
+export { StatsDisplay } from "./StatsDisplay"
